# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Local Google Classroom Dashboard", "version": "1.0.0" },
  "paths": {
...
    "/api/feedback/tickets/{ticket_id}": {
      "get": {
        "tags": ["feedback"],
        "summary": "Get My Ticket",
        "description": "One own ticket with its whole conversation, or 404.",
        "operationId": "get_my_ticket_api_feedback_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Ticket Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TicketDetailOut" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/feedback/tickets/{ticket_id}/messages": {
      "post": {
        "tags": ["feedback"],
        "summary": "Reply To Ticket",
        "description": "Reply to an own ticket.\n\nThe body carries ``body_markdown`` only. ``author_type``,\n``author_user_id``, ``author_email`` and ``display_name`` are not part of the\nmodel, so sending them changes nothing — a user cannot post under somebody\nelse's name (ADR-0035).\n\nReplying to a ``resolved`` ticket reopens it (``in_progress``), which the\nUI shows immediately.",
        "operationId": "reply_to_ticket_api_feedback_tickets__ticket_id__messages_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Ticket Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TicketDetailOut" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/feedback/attachments/{attachment_id}": {
      "get": {
        "tags": ["feedback"],
        "summary": "Download Attachment",
        "description": "Download an attachment of a ticket the caller OWNS (or any, as an admin).\n\nThe ownership check is a single joined statement: the attachment's ticket\nmust belong to the caller. Somebody else's file is 404 — indistinguishable\nfrom a missing one. Administrators may read any attachment, because support\nhas to see what was sent; the fallback path is still 404 for a missing row.",
        "operationId": "download_attachment_api_feedback_attachments__attachment_id__get",
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Attachment Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/admin/feedback/tickets": {
      "get": {
        "tags": ["admin"],
        "summary": "List Tickets",
        "description": "Every ticket, filtered by status/category/search, paginated.\n\n``q`` matches a subject or a message body as a substring (case-insensitive\nthrough ``ILIKE``; SQLite's ``like`` is case-insensitive for ASCII, which is\nwhat the tests assert). The filters are ANDed and the ordering is the same\n\"newest activity first\" as everywhere else, so the list stays stable while\nsomebody works on it.",
        "operationId": "list_tickets_api_admin_feedback_tickets_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Status"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Category"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "maxLength": 200 },
                { "type": "null" }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminTicketListOut" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/feedback/tickets/{ticket_id}": {
      "get": {
        "tags": ["admin"],
        "summary": "Read Ticket",
        "description": "Any ticket with its full conversation and the real author identities.\n\nNot filtered by the caller — an administrator is supposed to see every\nticket. The dependency above already answered 401/403.",
        "operationId": "read_ticket_api_admin_feedback_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Ticket Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTicketDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["admin"],
        "summary": "Set Status",
        "description": "Change the ticket status — and ONLY the status.\n\nThe body is ``{\"status\": \"resolved\"}``; anything else is ignored by the\nmodel, so this endpoint cannot become a way to rewrite a message, an author\nor an owner. An unknown status is 422 (the closed set).",
        "operationId": "set_status_api_admin_feedback_tickets__ticket_id__patch",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Ticket Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTicketDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "Delete Ticket",
        "description": "Delete a ticket permanently, with its conversation and its files.\n\nReal deletion, no ``deleted_at`` and no restore (ADR-0035): the row goes,\n``ON DELETE CASCADE`` takes the messages and attachment rows with it, and the\nfiles are unlinked from the volume. 204 on success, 404 for a ticket that\ndoes not exist — afterwards the id answers 404 for the owner as well, and\nthe ticket is gone from a fresh list.\n\nA file that cannot be unlinked does NOT fail the request: the row deletion\nis the contract, and the filesystem cleanup is best effort.",
        "operationId": "delete_ticket_api_admin_feedback_tickets__ticket_id__delete",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Ticket Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/admin/feedback/tickets/{ticket_id}/messages": {
      "post": {
        "tags": ["admin"],
        "summary": "Reply To Ticket",
        "description": "Answer a ticket under a chosen public display name.\n\nThe stored identity is split in two, and this is the whole point of the\nfeature:\n\n- ``author_user_id`` / ``author_email`` = the AUTHENTICATED administrator,\n  taken from the session — never from the request, so the audit trail stays\n  correct;\n- ``display_name`` = the public label the ticket owner sees (default\n  ``GoogleClassHelp Support``, capped at 100 characters). Two administrators\n  may therefore answer under different names while both are recorded\n  correctly internally.",
        "operationId": "reply_to_ticket_api_admin_feedback_tickets__ticket_id__messages_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Ticket Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTicketDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminMessageOut": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "author_type": { "type": "string", "title": "Author Type" },
          "display_name": { "type": "string", "title": "Display Name" },
          "body_markdown": { "type": "string", "title": "Body Markdown" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "attachments": {
            "items": { "$ref": "#/components/schemas/AttachmentOut" },
            "type": "array",
            "title": "Attachments"
          },
          "author_user_id": { "type": "integer", "title": "Author User Id" },
          "author_email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Author Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author_type",
          "display_name",
          "body_markdown",
          "created_at",
          "author_user_id"
        ],
        "title": "AdminMessageOut",
        "description": "A message with the internal identity an administrator may read."
      },
      "AdminTicketDetailOut": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "category": { "type": "string", "title": "Category" },
          "subject": { "type": "string", "title": "Subject" },
          "status": { "type": "string", "title": "Status" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "user_id": { "type": "integer", "title": "User Id" },
          "user_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "User Name"
          },
          "user_email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "User Email"
          },
          "messages": {
            "items": { "$ref": "#/components/schemas/AdminMessageOut" },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "subject",
          "status",
          "created_at",
          "updated_at",
          "user_id"
        ],
        "title": "AdminTicketDetailOut",
        "description": "Any ticket with the full conversation and the real author identities."
      },
      "AdminTicketListOut": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/AdminTicketOut" },
            "type": "array",
            "title": "Items"
          },
          "total": { "type": "integer", "title": "Total" },
          "limit": { "type": "integer", "title": "Limit" },
          "offset": { "type": "integer", "title": "Offset" }
        },
        "type": "object",
        "required": ["total", "limit", "offset"],
        "title": "AdminTicketListOut",
        "description": "A page of the admin list plus the total for the pagination control."
      },
      "AdminTicketOut": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "category": { "type": "string", "title": "Category" },
          "subject": { "type": "string", "title": "Subject" },
          "status": { "type": "string", "title": "Status" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count",
            "default": 0
          },
          "user_id": { "type": "integer", "title": "User Id" },
          "user_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "User Name"
          },
          "user_email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "User Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "subject",
          "status",
          "created_at",
          "updated_at",
          "user_id"
        ],
        "title": "AdminTicketOut",
        "description": "A ticket row for the admin list, with its owner."
      },
...
      "AttachmentOut": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "original_name": { "type": "string", "title": "Original Name" },
          "content_type": { "type": "string", "title": "Content Type" },
          "size_bytes": { "type": "integer", "title": "Size Bytes" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_name",
          "content_type",
          "size_bytes",
          "created_at"
        ],
        "title": "AttachmentOut",
        "description": "Metadata of one uploaded file; the bytes come from the download endpoint."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageOut": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "author_type": { "type": "string", "title": "Author Type" },
          "display_name": { "type": "string", "title": "Display Name" },
          "body_markdown": { "type": "string", "title": "Body Markdown" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "attachments": {
            "items": { "$ref": "#/components/schemas/AttachmentOut" },
            "type": "array",
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author_type",
          "display_name",
          "body_markdown",
          "created_at"
        ],
        "title": "MessageOut",
        "description": "One conversation message as a regular user may see it.\n\nNo ``author_user_id`` and no ``author_email``: a regular user sees the\nPUBLIC byline of the conversation and nothing else."
      },
...
      "TicketDetailOut": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "category": { "type": "string", "title": "Category" },
          "subject": { "type": "string", "title": "Subject" },
          "status": { "type": "string", "title": "Status" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "messages": {
            "items": { "$ref": "#/components/schemas/MessageOut" },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "subject",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "TicketDetailOut",
        "description": "One ticket with its whole conversation, as its OWNER sees it."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}